+++ src/compas_rhino/artists/networkartist.py
@@ -2,16 +2,12 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
 
-from functools import partial
 import compas_rhino
 
 from compas.geometry import centroid_points
-from compas.utilities import color_to_colordict
 
 from compas.artists import NetworkArtist
 from .artist import RhinoArtist
-
-colordict = partial(color_to_colordict, colorformat='rgb', normalize=False)
 
 
 class NetworkArtist(RhinoArtist, NetworkArtist):
@@ -174,14 +170,14 @@ class NetworkArtist(RhinoArtist, NetworkArtist):
 
         """
         self.node_color = color
-        node_xyz = self.node_xyz
         nodes = nodes or self.nodes
+        node_xyz = self.node_xyz
         points = []
         for node in nodes:
             points.append({
                 'pos': node_xyz[node],
                 'name': "{}.node.{}".format(self.network.name, node),
-                'color': self.node_color.get(node, self.default_nodecolor)
+                'color': self.node_color[node].rgb255
             })
         return compas_rhino.draw_points(points, layer=self.layer, clear=False, redraw=False)
 
@@ -204,14 +200,14 @@ class NetworkArtist(RhinoArtist, NetworkArtist):
 
         """
         self.edge_color = color
-        node_xyz = self.node_xyz
         edges = edges or self.edges
+        node_xyz = self.node_xyz
         lines = []
         for edge in edges:
             lines.append({
                 'start': node_xyz[edge[0]],
                 'end': node_xyz[edge[1]],
-                'color': self.edge_color.get(edge, self.default_edgecolor),
+                'color': self.edge_color[edge].rgb255,
                 'name': "{}.edge.{}-{}".format(self.network.name, *edge)
             })
         return compas_rhino.draw_lines(lines, layer=self.layer, clear=False, redraw=False)
@@ -220,7 +216,7 @@ class NetworkArtist(RhinoArtist, NetworkArtist):
     # draw labels
     # ==========================================================================
 
-    def draw_nodelabels(self, text=None, color=None):
+    def draw_nodelabels(self, text=None):
         """Draw labels for a selection nodes.
 
         Parameters
@@ -228,9 +224,6 @@ class NetworkArtist(RhinoArtist, NetworkArtist):
         text : dict[int, str], optional
             A dictionary of node labels as node-text pairs.
             The default value is None, in which case every node will be labelled with its key.
-        color : tuple[int, int, int] or dict[int, tuple[int, int, int]], optional
-            Color of the labels.
-            The default color is the same as the default color of the nodes.
 
         Returns
         -------
@@ -238,27 +231,19 @@ class NetworkArtist(RhinoArtist, NetworkArtist):
             The GUIDs of the created Rhino objects.
 
         """
-        if not text or text == 'key':
-            node_text = {node: str(node) for node in self.nodes}
-        elif text == 'index':
-            node_text = {node: str(index) for index, node in enumerate(self.nodes)}
-        elif isinstance(text, dict):
-            node_text = text
-        else:
-            raise NotImplementedError
+        self.node_text = text
         node_xyz = self.node_xyz
-        node_color = colordict(color, node_text.keys(), default=self.default_nodecolor)
         labels = []
-        for node in node_text:
+        for node in self.node_text:
             labels.append({
                 'pos': node_xyz[node],
                 'name': "{}.nodelabel.{}".format(self.network.name, node),
-                'color': node_color[node],
-                'text': node_text[node]
+                'color': self.node_color[node].rgb255,
+                'text': self.node_text[node]
             })
         return compas_rhino.draw_labels(labels, layer=self.layer, clear=False, redraw=False)
 
-    def draw_edgelabels(self, text=None, color=None):
+    def draw_edgelabels(self, text=None):
         """Draw labels for a selection of edges.
 
         Parameters
@@ -266,9 +251,6 @@ class NetworkArtist(RhinoArtist, NetworkArtist):
         text : dict[tuple[int, int], str], optional
             A dictionary of edgelabels as edge-text pairs.
             The default value is None, in which case every edge will be labelled with its key.
-        color : tuple[int, int, int] or dict[tuple[int, int], tuple[int, int, int]], optional
-            Color of the labels.
-            The default color is the same as the default color of the edges.
 
         Returns
         -------
@@ -276,20 +258,14 @@ class NetworkArtist(RhinoArtist, NetworkArtist):
             The GUIDs of the created Rhino objects.
 
         """
-        if text is None:
-            edge_text = {edge: "{}-{}".format(*edge) for edge in self.edges}
-        elif isinstance(text, dict):
-            edge_text = text
-        else:
-            raise NotImplementedError
+        self.edge_text = text
         node_xyz = self.node_xyz
-        edge_color = colordict(color, edge_text.keys(), default=self.default_edgecolor)
         labels = []
-        for edge in edge_text:
+        for edge in self.edge_text:
             labels.append({
                 'pos': centroid_points([node_xyz[edge[0]], node_xyz[edge[1]]]),
                 'name': "{}.edgelabel.{}-{}".format(self.network.name, *edge),
-                'color': edge_color[edge],
-                'text': edge_text[edge]
+                'color': self.edge_color[edge].rgb255,
+                'text': self.edge_text[edge]
             })
         return compas_rhino.draw_labels(labels, layer=self.layer, clear=False, redraw=False)